wms-demo/backend/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
sqlalchemy
httpx
passlib[bcrypt]
python-jose
reportlab
